backend/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
PyJWT
python-dotenv
pdfplumber
google-generativeai
pydantic
python-multipart